feat: use Collection as a resource directly instead of the CollectionResource
